+++ modules/API/CHANGEDB.php
@@ -6,10 +6,14 @@ Copyright © 2010, Gibbon Foundation
 Gibbon™, Gibbon Education Ltd. (Hong Kong)
 */
 
-//Updated to v1.3.04
+//Updated to v1.3.05
 $count = 0;
 
 ++$count;
 $sql[$count][0] = '1.3.04';
+$sql[$count][1] = '';
+
+++$count;
+$sql[$count][0] = '1.3.05';
 $sql[$count][1] = '';
 ?>